+++ lib/templates/project/manifest.rb
@@ -2,10 +2,12 @@
 stylesheet 'ie6.sass', :media => 'screen, projection', :condition => "lt IE 7"
 image 'button_bg.png'
 
+print "Check out http://brandonmathis.com/projects/fancy-buttons/guide/ to get a guide on how to use fancy-buttons"
+
 #print "Install example files? (html and buttons.sass) (Y/n)?" 
 #response = gets
 #if response.downcase == "y" || response.downcase == "\n"
-file 'index.html'
-stylesheet 'buttons.sass', :media => 'screen, projection'
+#file 'index.html'
+#stylesheet 'buttons.sass', :media => 'screen, projection'
 #end
 #print 'Don\'t forget to @import "fancy-buttons"'